Convert union/intersection speculation error messages to the new format
Summary:
This diff does a lot of stuff.

For union/intersection speculation error messages I wanted the same basic bullet-point structure as grouped error messages introduced in D6836971. So part of this diff is a refactor of the error rendering code to massage grouped errors into a `message_group`. A `message_group` is similar to a `<ul>` in HTML and can be nested with arbitrary levels of depth.

Then, I introduce a distinction between `Normal` error messages and `Speculation` error messages. `Speculation` error messages carry around all the error objects corresponding to failing speculation branches. In addition to carrying around the score! Importantly, with this diff we are no longer applying the score of speculation error messages in `Flow_js`. Instead we do so in `Errors` where hiding branches can be configured by the Flow client with the new `--show-all-branches` flag. (A user might want to see all branches for debugging purposes. However, this feature should only really ever be used by the Flow team.)

Before printing errors we take the distinct `Normal` or `Speculation` form and turn them into a `message_group`. This is done by the `message_group_of_error` function which is the largest new addition in this diff. This function applies a couple of important transformations:

1. Hides speculation branches with a low score and reports when it does so.
2. If there is one speculation branch with a higher score then all other branches, we show that branch alone instead of a list of branches.
3. Flattens nested speculation branches.

There are a couple of things to notice in the error format. We don't point to the union type //or// the branches which caused an error. We only point to the root location and the incompatibility locations. This is consistent with all our other error messages. We don't show every intermediate object type in an error with heavy nesting.

The only indication that you got a speculation error are the words "Either" and "Or" which begin each speculation error list item. Otherwise, speculation errors are pretty consistent with all of our other errors.

This consistency is a desirable property. We provide the user with the critical information necessary to quickly debug their programs in 80% of cases, while still providing enough information to prompt the user to dive deep into their program in the remaining challenging 20% of cases.

This diff does introduce a slight error message regression for intersection function types. Before, when we could use the score to choose a single error we would throw that one error directly. Now, instead, we generate the error in two different places. One for the top of the speculation and once for each branch. Then at render time, we put the two parts together. This means calling an overloaded function with an incorrect parameter will give a slightly different error message. You can see this happen a couple of times in the tests.

I'll be sending a diff very shortly after this one to clean up `Flow_error`.

Reviewed By: gabelevi

Differential Revision: D6934699

fbshipit-source-id: 96ec1876359116b8a522d29a10b9a29915f90db5