Add a reference to TypeArgCompatibility frame
Summary:
In D6669286 gabelevi points out some error messages which are a little hard to understand because we only print the text "in type argument `T`" without any extra context. "in type argument `T`" could mean there was an incompatibility in an array element or in some user defined type.

This diff does two things to improve `TypeArgCompatibility` frames in error messages:

1. Add a reference location to the type argument's definition. This allows the user to see file in which the type or class was defined.
2. Add an `ArrayElementCompatibility` to provide a custom error message for array element incompatibilities. (This was motivated by not easily being able to access the definition location for `Array<T>`, but is an improvement on its own.)

Reviewed By: gabelevi

Differential Revision: D6875505

fbshipit-source-id: ea7d3ecc933cb56235846f7268e0eaf7aaef2912