Replace "X, Y, and then Z" frame rendering with "Z of Y of X"
Summary:
In D6548360 fishythefish request trying to print `use_op` frame paths in "C of B of A" format instead of "A, B, and then C" format. I agree that the "and then" format isn't ideal. However, it was the best format I could come up with. Publishing this diff to explain why I ended up at the "and then" format and to expirement with alternatives.

We can think of incompatibility errors in 3 parts.

1. An "incompatibility." e.g. `string \!= number`.
2. An "operation." e.g. `X cannot be cast to Y`.
3. A series of "frames." e.g. "property `x`", "type argument `T`", or "the first argument".

The job of `unwrap_use_ops_friendly` is to take these 3 parts and render them in a human readable format.

Here is a diagram of a complex example error which we can propose different renderings for:

```
Operation: Op
Frame: A
Frame: B
Frame: C
Frame: D
Incompatibility: X \!= Y
```

The render format I chose before this diff was: "Cannot `Op` because in `A`, `B`, `C`, and then `D`; `X` is not compatible with `Y`."

This had the advantage of describing the error in the order which things happened. We started at the `Op`. Then we looked in `A`, `B`, `C`, and then `D`. At that point we found an incompatibility between `X` and `Y`.

As fishythefish describes, "and then seems to describe a process," which is correct. It describes the type-checkers progress. However, the user does not think of their program in terms of "type checking progress." The user thinks about their program in terms of data structures. `C` is inside `B`, `B` is inside `A`, and so on.

This diff changes the render format to something which may better reflect the user's thought process: "Cannot `Op` because in `D` of `C` of `B` of `A`, `X` is not compatible with `Y`."

This ends up looking ok in a lot of our test suite. Especially for the errors where we render the location of `X` instead of the location of `Op`.

Despite some messages looking better, there are a couple reasons why I don't like this:

- We render the error frames backwards. The logical order of the error is `Op` -> `A` -> `B` -> `C` -> `D` -> `X \!= Y`. However, we flip the order of `A` through `D`. While saying "`D` of `B` of `C` of `A`" makes more sense given the user thinks about their program in data structures, it does not make sense when surrounded by `Op` and `X \!= Y`. I don't want to flip the order of `Op` and `X \!= Y`, because in a CLI renderer rewrite we will dedupe errors that share the same `Op`.
- We put a conjunction in between every frame. Punctuation makes it easier to see each frame individually. Take for instance: "in type argument `Yield` of the return value of property `@iterator`." There are 3 frames here. It is easier to identify these frames when we render: "in property `@iterator`, the return value, and then type argument `Yield`."
- Successive object properties are flattened into a single path frame. e.g. "property `a.b.c`" instead of "property `a`, property `b`, and then property `c`". This behavior combined with reversing the frames can lead to messages with inconsistent frame ordering. e.g. "`D` of `B.C` of `A`." This doesn't matter to users, though, since users can't know how many frame data structures went into creating their message.

I agree with fishythefish that there may be a better way to render frames then using "and then" in a conjunction list. However, I don't like the first obvious idea (which I had as well) of reversing the list and combining frames with "of".

I'm not convinced this diff is a good idea. If you accept I may not land. I would prefer that we discover a better solution without the tradeoffs of the reversed "of" list.

Putting this up as a "needs review" diff to get feedback.

@[4811713:glevi] commented a good suggestion for solving the ordering problem. Use possesion. So we would render “A’s B’s C’s D.” It’s an open question as to whether we could programmatically implement the correct gramatical rules so that this always works.

@[4802170:leebyron] proposed another suggestion which is what I’ve updated this diff with. Lee makes the insight that the incompatibility (`X != Y`) is what the user cares about. Much more then the frames it took to get there. So Lee suggested putting the frames //at the end of the message//. e.g. “Cannot `Op` because `X != Y` in `D` of `C` of `B` of `A`.” This allows us to use the “of” chaining @[100001558549237:fishythefish] suggested which is more natural to read.

Flipping the order especially makes sense for grouped error messages. Before it would be common to have an error message like:

```
Cannot Op because:
- in property `A` ...
- in property `A` ...
- in property `A` ...
- in property `A` ...
- in property `A` ...
```

All of the sub-errors starting with “in property `A`” made it easy to tune out the repeated stuff.

In regards to my prior concerns, I’m much less worried about flipping the frame order if we put it at the end. If anything flipping the frame order makes more sense when it is at the end since we lead from `X != Y` back to `Op`.

The successive property flattening could still be an issue. Let me know if you want it removed. It doesn’t look too out of place in the test suite although I’m sure we could produce a patholoigical example where that behavior is undesireable.

Reviewed By: gabelevi

Differential Revision: D6669283

fbshipit-source-id: 11bc02c27c492c48cccb1b15bff68f46e438565d