Fix `canonicalize_filenames` for non-existent files
Summary:
When I turned off file-watching in `./tool test`, OSX tests
started to fail. When I looked at the logs, I saw the following:

After adding a file:

```
[2018-02-01 15:09:14] Request: force-recheck /private/var/folders/m4/7j4c7dkj16qfxhtn_96p0flchhy4jh/T/flow/tests/2e4e792091/idezSlazy/6/dependency.js (focus = false)
```

After removing a file:

```
[2018-02-01 15:09:14] Request: force-recheck /var/folders/m4/7j4c7dkj16qfxhtn_96p0flchhy4jh/T/flow/tests/2e4e792091/idezSlazy/6/dependency.js (focus = false)
```

It turns out `/var` is a symlink to `/var/folders` and that `realpath`
basically fails for non-existent files, so we never properly
canonicalize the removed file.

We could rely on the caller to get this right, but I think it's better
if Flow does its best to figure out which file was removed.

Reviewed By: samwgoldman

Differential Revision: D6875030

fbshipit-source-id: 41f8db2809219b6cef271572437903a0ad4f9136