Update `./tool test` to use `--file-watcher none`
Summary:
Like `./runtests.sh`, we need to use `force-recheck` during a
test to ensure that the test script doesn't race the file watcher.
Turning off the file watcher guarantees that we will never race the file
watcher.

Another benefit is that the ide tests no longer notice the double
recheck that happens thanks to the file watcher and the force-recheck
both reporting the same change.

Reviewed By: samwgoldman

Differential Revision: D6761260

fbshipit-source-id: dcf6698152aa546397e39f1374ab904986ad63f9