Add `--profile` flag to `flow force-recheck`
Summary:
If you run a Flow server with `--profile`, it will print
profiling info after each recheck to the .log file, but that data is
hard to extract.

A better option is to add `flow force-recheck --profile`. Normally,
`flow force-recheck` exits immediately, but with `--profile` it will
wait for the recheck to finish and then print the stats for the recheck.

I've also added the `--json` and `--pretty` flags, to format the output.

Reviewed By: samwgoldman

Differential Revision: D6761984

fbshipit-source-id: dd901f35c654d2a0a5cc742e158103026caefca8