Fix `--retry-if-init false`
Summary:
Many of our client commands have a flag called
`--retry-if-init`, to which you can pass `true` or `false`. The default
is `true`.

The intention of this flag is for a client command to exit immediately
when the server is still initializing. This functionality broke when I
rewrote the connection code for the Flow server monitor.

This diff restores the behavior, with one minor difference. The exit
code returned is Out_of_retries (7), whereas before it was
Server_initializing (1) which no longer exists.

Reviewed By: samwgoldman

Differential Revision: D6763299

fbshipit-source-id: d53cd406ad9cb1b26751f6f744b879395cb3e4ce