ssa builder bug around do/while + if + continue + return
Summary:
A fairly tricky bug that was exposed by a combination of control-flow constructs, but I'll try to
explain anyway.

Suppose that we have some straight-line code that exits abnormally, like `... return;`. Then we not
only record the `return` as an outgoing control-flow edge, but also raise an OCaml exception to
simulate what happens at run time...effectively considering the code that follows the `return` dead,
which improves the precision of SSA.

On the other hand, suppose that there is some branch that exits abnormally, like `if (...) {
... continue; }`. Then, as above, we record the `continue` as an outgoing control-flow edge. But we
must not raise an OCaml exception at the merge point after the branch, since we do not know whether
that branch will be taken at run time.

Now combine these two: `if (...) { ... continue; } ... return;` We would then have two outgoing
edges, for `continue` and `return`, but only one OCaml exception, for `return`.

Finally, consider an enclosing `while(...) { ... }` that forms the target of the `continue` edge
outgoing from somewhere in the middle of the loop body. Since we have an OCaml exception for the
`return` at the end of the loop body, we can assume that all control-flow paths reaching the end
abnormally exit with `return`. But we were mistakenly extending this assumption when going back to
the loop check: the `continue` edge as well as the `return` edge have to merge before the loop
check, and we were not merging the `continue` edge when the `return` edge existed.

Ultimately this caused us to mistakenly think that a write before the `continue` would not reach a
read in the loop check, because we would consider the read dead due to the `return` in the middle.

Reviewed By: mroch

Differential Revision: D7019599

fbshipit-source-id: 02c18644c24b2f74051465c0011320e8ecb76622