Evaluate type of `super` all the time when it has a property accessed on it
Summary:
I'm about to put up a change that includes in the type table the type of the receiver for property accesses. This change is necessary to both (a) include that type, and (b) set the type in the type table (as is done currently) whether or not we find a refinement.

This matches the behavior of the case for property accesses on any other objects (just below this one).

Reviewed By: samwgoldman

Differential Revision: D7050648

fbshipit-source-id: c15e4d600c1bc616075499a651de8d443634fa11