Lazy mode message for flow status
Summary:
`flow status` reports all the known errors. This means it can
be a little tricky to use with a server that is in lazy mode, since
"No errors" really means "No errors in the subset of the graph which
lazy mode checked".

One solution is to add a message to `flow status` alerting the user that
the server is in lazy mode and linking to documentation.

This diff also adds that documentation

Reviewed By: avikchaudhuri

Differential Revision: D7047496

fbshipit-source-id: ce02b52f607e207ba2d700c82bb3cf34ab43b6e6