Use TypeDestructorTrigger to resolve AnnotT instead of EvalDestructor
Summary:
Using the choice kit concretization machinery attempts to "fully" resolve the
underlying annot type. This full type resolution for classes might run into
OpenUnresolved tvars -- e.g., unannotated return types or class properties --
which get scrubbed out to any.

We don't need to fully resolve these annots, though; we just need to resolve
the top-level type. This diff re-uses the TypeDestructorTriggerT type to wait
for the annot's lower bound before trying to evaluate the destructor.

Annots expect to be repositioned. Since TypeDestructorT is very high up in the
__flow pattern match, I didn't want to rely on standard repositioning
constraints. Instead, I added an optional repositioning argument to
TypeDestructorTriggerT, which only matters when it appears as an upper bound.

Finally, because AnnotT is no longer handled in eval_destructor, I needed to
add a special case for TypeDestructorTriggerT lower bounds of an AnnotT upper
bound, to ensure we resolve the lower bound of the AnnotT. This is essentially
a combined version of `(_, UseT (_, AnnotT _))` and the
`(TypeDestructorTriggerT _, UseT _)` rules.

Reviewed By: calebmer

Differential Revision: D6953539

fbshipit-source-id: 468b7bc75d6dbafb2d8ded08b908bc16129b0c22