Make Marshal_tools.to_fd_with_preamble return payload size
Summary:
It's pretty common for write APIs to return the bytes written, so I think this is a good change, even if almost every caller ignores the return value.

I intend to use Marshal_tools to replace the custom-rolled marshaling code which the workers use. We complain if the workers are marshaling too much data, so I need this change for that.

Reviewed By: alexchow

Differential Revision: D6846316

fbshipit-source-id: ef3ed607c043975089ff8a129c702ec48aaa61d9