disjoint union representation optimization
Summary:
A disjoint union of objects can be represented as a map from the literal values of a common sentinel
property to the corresponding cases of the union. Such a representation speeds up membership checks,
avoiding speculation altogether.

Since it is not given what a common sentinel property might be, the representation keeps around all
possible ones.

Reviewed By: samwgoldman

Differential Revision: D7110034

fbshipit-source-id: 26bf23efed465bbea25f48e6066a9e6872e59836