[hack] Add a 60 second timeout check to the loop in hh_shared.c
Summary:
While we should figure out some more general way to cleanup after ourselves and terminate all the processes no matter where are they stuck, the practice shows that all of them are stuck in this particular loop, and we can easily prevent it.

I didn't see it affect init time, and this is in-line with my expectations. The optimistic path that majority of workers will hit majority of the time is not modified.

(Note: this ignores all push blocking failures!)

Reviewed By: gabelevi

Differential Revision: D7120676

fbshipit-source-id: 1a339d30b8406e0323e927e0ba927d46f23cefa6