Optional chaining lexer/parser support
Summary:
This adds the `T_PLING_PERIOD` token (following the existing naming convention) to represent the optional chaining operator (`?.`) and implements lexing and parsing rules for it, per the [TC39 proposal](https://github.com/tc39/proposal-optional-chaining) and associated [Babylon plugin](https://github.com/babel/babylon/pull/545).

This includes a lookahead in the lexer to ensure that expressions like `1?.2:3` are read as `1 ? .2 : 3` (using the ternary operator) for backward compatibility, as mandated in the proposal.

At the AST/parser level, this adds an `optional` flag to `Call`, `Member`, and `New` expression nodes, mirroring the Babylon implementation. The accompanying test suite matches the expectations set by the Babylon plugin's tests.

Support for this feature is controlled by the `esproposal_optional_chaining` flag, which can also be set via `.flowconfig`.

Reviewed By: gabelevi

Differential Revision: D7029173

fbshipit-source-id: 5cb657c78f343348a9d3bbde09db837441cb7dec