Clear envs from context before serializing
Summary:
We need to keep the scope records around after local inference into merge,
because a call might be deferred due to latent dependencies.

However, once a context is fully merged, there is never any reason to keep the
scope records around. Before, we were discarding most of the data in context
optimization.

Reviewed By: avikchaudhuri

Differential Revision: D7132031

fbshipit-source-id: b7124208569393fce313748555292c3f2c3cbd6d