Add a test.log file to tool tests
Summary:
While investigating some ide test timeouts, I was curious about
the times at which the tool test was sending IDE notifications to the
server. So I decided to add a log. That way I could compare

 /tmp/flow/tests/cc4229b978/idezSlazy/tmp/1/test.log
 /tmp/flow/tests/cc4229b978/idezSlazy/tmp/1/zStmpzSflowzStestszScc4229b978zSidezZSlazZyzS1.monitor_log

To see what the test runner was logging and what the server monitor was
logging.

Reviewed By: nmote

Differential Revision: D7113581

fbshipit-source-id: 962a232f0f2692bb48459dc347376bfb0fbea257