Error on prop lookups in conditionals that always fail.
Summary:
Currently, Flow doesn't complain about things like

  if (obj.tyypo) { ... }

because `obj` can be some complicated type and it's hard to prove that
`obj` never has the property `tyypo`. For example, it could be a union
type, and one of the branches might have that property.

This diff proposes a solution where we keep track of whether the prop
lookup ever succeeds. If it always misses, then we raise an error

Reviewed By: samwgoldman

Differential Revision: D6993151

fbshipit-source-id: 7d9e0813891d3b3100d3b6fcaccd0d8c120399d5