Log when server is killed by a signal
Summary:
Currently the monitor watches for when a server dies. Based on
the exit code, it decides whether to start a new server or also die. It
notices when the server died due to a signal, and always tries to start
a new server after a signal.

However, servers that died due to a signal lack `EXIT` event logging.
This tries to fix it by logging the event from the monitor process

Reviewed By: nmote

Differential Revision: D7124417

fbshipit-source-id: 2dfccb5acc87301be08d8f8c30f231f930d122d1