get rid of early flattening
Summary:
We used to flatten / finally optimize unions when making new unions via ident_map. This was
obviously non-ideal since tvars may not be resolved at that time, leading to lost opportunities for
optimization later when they *were* resolved. However without this, recursion limit errors were
possible in cases where, because of unresolved tvars, we would change the optimization state
uselessly and that would cause new unresolved tvars to be created, and this would go on in a loop.

We can avoid this by being careful about not switching to a final optimization state when there are
unresolved tvars after flattening.

As a result, we now have final optimization states that are idempotent, as in attempting to optimize
again will not change the state. This should be possible to exploit for further optimization.

Reviewed By: samwgoldman

Differential Revision: D7118062

fbshipit-source-id: 114dc563bfc9446b13704255240a8f87a43b9625