clear TestPropHits whenever various flow caches are cleared
Summary:
A recent diff caused the number of errors reported to be nondeterministic on large codebases. This
diff seems to fix the issue.

But how? I remember that clearing the various flow caches at the beginning, rather than at the end,
of `infer_ast` is necessary, but I don't quite remember why, so I'll try guessing.

Perhaps because stuff could happen between `infer_ast` calls, which could leave junk that would lead
to a bad initial state. If this stuff happened in a nondeterministic order, that might explain the
overall nondeterminism.

Indeed `merge_strict_component` calls `implicit_require_strict` + `infer_ast` for each file in a
component, then clears the various flow caches and calls `explicit_*_require_strict`. This can leave
the memory of a worker in a pretty weird state, and when that worker processes another component, oh
well.

Reviewed By: gabelevi

Differential Revision: D7166990

fbshipit-source-id: 59284557cf27646c32beac211b11fe4a26537442