Fix GC for shared sig_cxs
Summary:
In D7135469 I changed the way contexts are laid out. Now, all contexts in a
cycle share a single graph, property maps index, etc.

That change broke GC in a pretty trivial way, which this diff fixes.

The Gc_js module is essentially mark-and-sweep, except that it's sensitive to
polarities as well. The polarities are not important for this bug or fix, so
just consider a normal mark and sweep GC.

For every context in a component, we walk the type graph from exports, marking
every tvar we see. Any tvar we didn't mark in that step is removed from the
graph.

That works fine when every context has a separate graph, but when they all
share a graph, later GC sweeps were clearing out marked tvars from previous
files.

This fix keeps the marked set between files, so we don't clear those tvars.

Reviewed By: gabelevi

Differential Revision: D7206334

fbshipit-source-id: f7984a118dcb92f73aafd0619b232c1a4b8fd298