[Normalizer] Distinguish recursive tvars from polymorphic tvars
Summary:
There are two flavors of type variables that `Ty.t` needs to know about:
* *Recursive* type variables (introduced during the resolution of any tvar in case it
is a recursive type). We use an integer as the identifiers for these.
* *Generic* type parameters. These are represented as strings, and correspond
to the original name of the type parameter.

Before, we were treating the later kind was treated as a named type. We now
distinguish this case and introduce the `TParam` constructor.

Reviewed By: samwgoldman

Differential Revision: D7266415

fbshipit-source-id: 5b25e745d519d05e81e43fb49029f18e1f1191d2