[Normalizer] Add provenance information to symbols of `Ty.t`
Summary:
The results produced by the normalizer often contain names that may originate
from a number of places:
(a) the current file,
(b) a library file,
(c) another file in the project, or
(d) a set of built-in names.

The case of (c) is interesting in that these names can be imported (c1) in the current
file or not (c2).

One of the normalizer's eventual goals is that the names it produces are in scope in the
current module. Several services that will depend on the normalizer will need this
information (e.g. gen-flow-files). Going back to the (c2) case from above, such names
should be banned from the output.

This commit sets the foundation for building a well-formedness invariant for the output
of the normalizer. It does this by adding //provenance// information for each name
appearing in a `Ty.t`.

To decide if the various names are properly imported, the normalizer's state is
initialized with information about the module's imported types.

Reviewed By: samwgoldman

Differential Revision: D7267988

fbshipit-source-id: 1f57c55b9160ad8d14c79d54c9f868eeded43c33