[hack] Add completionItem/resolve support to Hack LSP.
Summary:
The LSP has a concept of `completionItem/resolve`, which is a request that can add additional information to an autocomplete request. Now that Nuclide can support sending `completionItem/resolve` requests, the Hack LSP server should support them as well.

In this change, the only additional functionality we support in `completionItem/resolve` is filling in doc blocks.

Reviewed By: arxanas

Differential Revision: D7283031

fbshipit-source-id: de2c03a92eae75e49c976b2e6e05fd427f90cc64