'flow lsp' should forward LSP requests to server
Summary:
This diff allows the persistent client to forward LSP messages over to the server, and for the server to send LSP packets back.

It puts in just a placeholder handler on the server side, for get-definition. We'll be expanding this in the future.

Note the way logging works for persistent requests:

1. Hh_logger.debug "Persistent request: %s"
2. MonitorRPC.status_update Handling_request_start & <do the work> & MonitorRPC.status_update Finishing_up
3. OK? -> FlowEventLogger.persistent_command_success
4. ERROR? -> FlowEventLogger.persistent_command_success & LSP error message to client
5. EXCEPTION? -> Hh_logger.error "Uncaught" & FlowEventLogger.persistent_command_failure & LSP error message to client

Note that "persistent_command_success" doesn't mean that the command succeeded; it merely means that there weren't any logic bugs (i.e. exceptions) inside the handler; "persistent_command_failure" means there was an exception.

So: when we want to review persistent request issues in our logs, we have to (1) look for "success" to find all the ones whose data column has "Error", and (2) look for "failure".

Reviewed By: gabelevi

Differential Revision: D6508170

fbshipit-source-id: 625847f482655ea5e336760e7b974ad02786e6d5