kill some dead code in JSX typing
Summary:
D5395653 added a pattern for `Identifier _ when jsx_mode = None`, which makes the `if jsx_mode = None` branch in the following `Identifier _` pattern dead. I think this logic has subsequently moved to React_kit.

Now the only difference between the two branches is reasons, and a StrT vs a SingletonStrT. I bet these could be further combined but I'm not very familiar with this area.

Reviewed By: gabelevi

Differential Revision: D7291578

fbshipit-source-id: 1a58f6b83fa494fede4603d0649815086f9c9513