Use def_loc_of_reason for symbol provenance
Summary:
Before, the normalizer used `Reason.loc_of_reason` to compute the
`provenance` of a `Ty.symbol`. This was problematic since this approach was
not robust w.r.t. reason repositioning, and so the normalizer reported the latest
instance of a type as the provenance location instead of the actual definition location.

Now the normalizer uses `Reason.def_loc_of_reason` instead, to get the location
of the definition.

Reviewed By: samwgoldman

Differential Revision: D7314418

fbshipit-source-id: 57be9af982945690cc15378a8e8c53868f0fcef1