[flow-parser-bin] make `npm install` work locally without github
Summary:
before, running `npm install` downloaded the prebuilt libflowparser.a from github. this is a problem when trying to test a pre-release.

this diff introduces `make local`, which locally compiles libflowparser.a for your platform and uses that to build the node extension; it also introduces `make dist`, which has the previous behavior of downloading the prebuilt binaries from the release, for both linux and osx. these are bundled with the npm release so that you don't need ocaml to build the extension.

so, if you do `npm install` on a source checkout, it'll build libflowparser from source. if you install it from npm, it'll use the included libflowparser.a.

Reviewed By: gabelevi

Differential Revision: D7091623

fbshipit-source-id: 73eb24d4550df1fb6541708f059558d341f22287