Error on non-top-level CJS exports
Summary:
I'm moving towards stricter treatment of CommonJS exports, with the eventual goal of making `File_sig` the source of truth for exports, rather than relying on the typechecker to determine what symbol(s) a file exports.

To begin with, I am issuing an error for CJS exports that do not appear at the top level. I plan to add errors for other sketchy uses of `module` and `exports` but this diff is big enough already so I will do that as a follow-up.

Right now these errors are behind a config option, but eventually that config option will go away, so for Flow users, it would be a good idea to turn it on and address the issues it exposes sooner rather than later.

Reviewed By: samwgoldman

Differential Revision: D7278143

fbshipit-source-id: afc44b8112190a28076a8dc83f4695bf676f8619