[Easy] Don't say we're ready until after garbage collecting
Summary:
We used to send the "Ready" event immediately before garbage
collecting. That's kind of dumb. We should say we're ready after garbage
collecting, when we're actually ready to handle a request.

This is a bigger deal when I move file watching to the monitor process,
since a ready server will just stop and wait for a message from the
monitor.

Reviewed By: samwgoldman

Differential Revision: D7336759

fbshipit-source-id: 1edaf89bcc8aa6a317ad9cc7e1cf9e6b2e0129b7