[Easy] LwtLoop shouldn't catch all exceptions
Summary:
Uncaught exceptions from inside LwtLoop loops should not be
caught. They should propagate up to Lwt's async exception handler, which
will cause them to be logged and the process to exit.

The one exception is the Cancelled exception, which should be caught
since cancelling loops is an expected behavior

Reviewed By: samwgoldman

Differential Revision: D7336839

fbshipit-source-id: 415b6533bd954e77b848a5176af10e466a0bd0ef