refactor JSX title typing
Summary:
there are 3 different types of JSX calls (react, custom jsx and csx), and in some cases we special-case capitalized identifiers. this cleans up the various permutations so that it's easier to follow.

the only change in this diff is that the reason of non-capitalized JSX-pragma elements change from "`div`" to "JSX element `div`", which matches the capitalized reason, "JSX element `Foo`".

there's potential to further unify the react and JSX pragma cases, if it's OK to change from a StrT to a SingletonStrT in the latter.

Reviewed By: samwgoldman

Differential Revision: D7383005

fbshipit-source-id: 29a0a81a6826ee61f0208f62cc2c8ac1a8a0ffa2