[Easy] LwtInit.run_lwt
Summary:
Previously you needed two calls: one to set the engine and then
one to run the lwt thread. This combines them into a single call.

I originally built this because I was adding a third call. I eventually
didn't add that third call, but still liked this idea better than what I
had before. It also helps that xhpOfReactCommand.ml forgot to set the
engine, which shows that this is useful.

Reviewed By: samwgoldman

Differential Revision: D7272136

fbshipit-source-id: 05c09ee42bc91b8ec169838318104ef2b7146059